// src/generated/model/hentKravdetaljerJsonResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Tilbakekreving API
 * OpenAPI spec version: 1.0
 */
import type { HentKravdetaljerJsonResponseAvvik } from './hentKravdetaljerJsonResponseAvvik';
import type { HentKravdetaljerJsonResponseKrav } from './hentKravdetaljerJsonResponseKrav';
import type { HentKravdetaljerJsonResponseOppdragsgiver } from './hentKravdetaljerJsonResponseOppdragsgiver';
import type { HentKravdetaljerJsonResponseSkyldner } from './hentKravdetaljerJsonResponseSkyldner';

export interface HentKravdetaljerJsonResponse {
  krav: HentKravdetaljerJsonResponseKrav;
  oppdragsgiver: HentKravdetaljerJsonResponseOppdragsgiver;
  skyldner: HentKravdetaljerJsonResponseSkyldner;
  avvik?: HentKravdetaljerJsonResponseAvvik;
}